refactor: extract auth from shared
