--- src/shared/auth/index.ts
@@ -1,4 +1,0 @@
-export type { AuthSession, AuthUser, CurrentUserState } from '../types/auth-identity'
-export { getCurrentSession } from './state/get-current-session'
-export { getCurrentUserState } from './state/get-current-user'
-export { watchAuthState } from './state/on-auth-state-change'
